+++ server/tlsremote/tlsremote_test.go
@@ -7,14 +7,17 @@ import (
 	"fmt"
 	"io"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
 	"github.com/WatchBeam/clock"
 	"github.com/kolide/fleet/server/config"
 	hostctx "github.com/kolide/fleet/server/contexts/host"
+	"github.com/kolide/fleet/server/contexts/viewer"
 	"github.com/kolide/fleet/server/datastore/inmem"
 	"github.com/kolide/fleet/server/kolide"
+	"github.com/kolide/fleet/server/pubsub"
 	"github.com/kolide/fleet/server/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -291,6 +294,588 @@ func TestHostDetailQueries(t *testing.T) {
 			strings.HasPrefix(name, hostDetailQueryPrefix),
 			fmt.Sprintf("%s not prefixed with %s", name, hostDetailQueryPrefix),
 		)
+	}
+}
+
+func TestDetailQueriesWithEmptyStrings(t *testing.T) {
+	ds, svc, mockClock := setupOsqueryTests(t)
+	ctx := context.Background()
+
+	nodeKey, err := svc.EnrollAgent(ctx, "", "host123")
+	assert.Nil(t, err)
+
+	host, err := ds.AuthenticateHost(nodeKey)
+	require.Nil(t, err)
+
+	ctx = hostctx.NewContext(ctx, *host)
+
+	// With a new host, we should get the detail queries (and accelerated
+	// queries)
+	queries, acc, err := svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, len(detailQueries))
+	assert.NotZero(t, acc)
+
+	resultJSON := `
+{
+"kolide_detail_query_network_interface": [
+		{
+				"address": "192.168.0.1",
+				"broadcast": "192.168.0.255",
+				"ibytes": "",
+				"ierrors": "",
+				"interface": "en0",
+				"ipackets": "25698094",
+				"last_change": "1474233476",
+				"mac": "5f:3d:4b:10:25:82",
+				"mask": "255.255.255.0",
+				"metric": "",
+				"mtu": "",
+				"obytes": "",
+				"oerrors": "",
+				"opackets": "",
+				"point_to_point": "",
+				"type": ""
+		}
+],
+"kolide_detail_query_os_version": [
+		{
+				"platform": "darwin",
+				"build": "15G1004",
+				"major": "10",
+				"minor": "10",
+				"name": "Mac OS X",
+				"patch": "6"
+		}
+],
+"kolide_detail_query_osquery_info": [
+		{
+				"build_distro": "10.10",
+				"build_platform": "darwin",
+				"config_hash": "3c6e4537c4d0eb71a7c6dda19d",
+				"config_valid": "1",
+				"extensions": "active",
+				"pid": "38113",
+				"start_time": "1475603155",
+				"version": "1.8.2",
+				"watcher": "38112"
+		}
+],
+"kolide_detail_query_system_info": [
+		{
+				"computer_name": "computer",
+				"cpu_brand": "Intel(R) Core(TM) i7-4770HQ CPU @ 2.20GHz",
+				"cpu_logical_cores": "8",
+				"cpu_physical_cores": "4",
+				"cpu_subtype": "Intel x86-64h Haswell",
+				"cpu_type": "x86_64h",
+				"hardware_model": "MacBookPro11,4",
+				"hardware_serial": "ABCDEFGH",
+				"hardware_vendor": "Apple Inc.",
+				"hardware_version": "1.0",
+				"hostname": "computer.local",
+				"physical_memory": "17179869184",
+				"uuid": "uuid"
+		}
+],
+"kolide_detail_query_uptime": [
+		{
+				"days": "20",
+				"hours": "0",
+				"minutes": "48",
+				"seconds": "13",
+				"total_seconds": "1730893"
+		}
+],
+"kolide_detail_query_osquery_flags": [
+		{
+			"name":"config_tls_refresh",
+			"value":""
+		},
+		{
+			"name":"distributed_interval",
+			"value":""
+		},
+		{
+			"name":"logger_tls_period",
+			"value":""
+		}
+]
+}
+`
+
+	var results kolide.OsqueryDistributedQueryResults
+	err = json.Unmarshal([]byte(resultJSON), &results)
+	require.Nil(t, err)
+
+	// Verify that results are ingested properly
+	svc.SubmitDistributedQueryResults(ctx, results, map[string]string{})
+
+	// Make sure the result saved to the datastore
+	host, err = ds.AuthenticateHost(nodeKey)
+	require.Nil(t, err)
+
+	// osquery_info
+	assert.Equal(t, "darwin", host.Platform)
+	assert.Equal(t, "1.8.2", host.OsqueryVersion)
+
+	// system_info
+	assert.Equal(t, 17179869184, host.PhysicalMemory)
+	assert.Equal(t, "computer.local", host.HostName)
+	assert.Equal(t, "uuid", host.UUID)
+
+	// os_version
+	assert.Equal(t, "Mac OS X 10.10.6", host.OSVersion)
+
+	// uptime
+	assert.Equal(t, 1730893*time.Second, host.Uptime)
+
+	// osquery_flags
+	assert.Equal(t, uint(0), host.ConfigTLSRefresh)
+	assert.Equal(t, uint(0), host.DistributedInterval)
+	assert.Equal(t, uint(0), host.LoggerTLSPeriod)
+
+	mockClock.AddTime(1 * time.Minute)
+
+	// Now no detail queries should be required
+	host, err = ds.AuthenticateHost(nodeKey)
+	require.Nil(t, err)
+	ctx = hostctx.NewContext(ctx, *host)
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, 0)
+	assert.Zero(t, acc)
+
+	// Advance clock and queries should exist again
+	mockClock.AddTime(1*time.Hour + 1*time.Minute)
+
+	err = svc.SubmitDistributedQueryResults(ctx, kolide.OsqueryDistributedQueryResults{}, map[string]string{})
+	require.Nil(t, err)
+	host, err = ds.AuthenticateHost(nodeKey)
+	require.Nil(t, err)
+
+	ctx = hostctx.NewContext(ctx, *host)
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, len(detailQueries))
+	assert.Zero(t, acc)
+}
+
+func TestDetailQueries(t *testing.T) {
+	ds, svc, mockClock := setupOsqueryTests(t)
+	ctx := context.Background()
+
+	nodeKey, err := svc.EnrollAgent(ctx, "", "host123")
+	assert.Nil(t, err)
+
+	host, err := ds.AuthenticateHost(nodeKey)
+	require.Nil(t, err)
+
+	ctx = hostctx.NewContext(ctx, *host)
+
+	// With a new host, we should get the detail queries (and accelerated
+	// queries)
+	queries, acc, err := svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, len(detailQueries))
+	assert.NotZero(t, acc)
+
+	resultJSON := `
+{
+"kolide_detail_query_network_interface": [
+    {
+        "address": "192.168.0.1",
+        "broadcast": "192.168.0.255",
+        "ibytes": "1601207629",
+        "ierrors": "314179",
+        "interface": "en0",
+        "ipackets": "25698094",
+        "last_change": "1474233476",
+        "mac": "5f:3d:4b:10:25:82",
+        "mask": "255.255.255.0",
+        "metric": "1",
+        "mtu": "1453",
+        "obytes": "2607283152",
+        "oerrors": "101010",
+        "opackets": "12264603",
+        "point_to_point": "",
+        "type": "6"
+    }
+],
+"kolide_detail_query_os_version": [
+    {
+        "platform": "darwin",
+        "build": "15G1004",
+        "major": "10",
+        "minor": "10",
+        "name": "Mac OS X",
+        "patch": "6"
+    }
+],
+"kolide_detail_query_osquery_info": [
+    {
+        "build_distro": "10.10",
+        "build_platform": "darwin",
+        "config_hash": "3c6e4537c4d0eb71a7c6dda19d",
+        "config_valid": "1",
+        "extensions": "active",
+        "pid": "38113",
+        "start_time": "1475603155",
+        "version": "1.8.2",
+        "watcher": "38112"
+    }
+],
+"kolide_detail_query_system_info": [
+    {
+        "computer_name": "computer",
+        "cpu_brand": "Intel(R) Core(TM) i7-4770HQ CPU @ 2.20GHz",
+        "cpu_logical_cores": "8",
+        "cpu_physical_cores": "4",
+        "cpu_subtype": "Intel x86-64h Haswell",
+        "cpu_type": "x86_64h",
+        "hardware_model": "MacBookPro11,4",
+        "hardware_serial": "ABCDEFGH",
+        "hardware_vendor": "Apple Inc.",
+        "hardware_version": "1.0",
+        "hostname": "computer.local",
+        "physical_memory": "17179869184",
+        "uuid": "uuid"
+    }
+],
+"kolide_detail_query_uptime": [
+    {
+        "days": "20",
+        "hours": "0",
+        "minutes": "48",
+        "seconds": "13",
+        "total_seconds": "1730893"
+    }
+],
+"kolide_detail_query_osquery_flags": [
+    {
+      "name":"config_tls_refresh",
+      "value":"10"
+    },
+    {
+      "name":"config_refresh",
+      "value":"9"
+    },
+    {
+      "name":"distributed_interval",
+      "value":"5"
+    },
+    {
+      "name":"logger_tls_period",
+      "value":"60"
+    }
+]
+}
+`
+
+	var results kolide.OsqueryDistributedQueryResults
+	err = json.Unmarshal([]byte(resultJSON), &results)
+	require.Nil(t, err)
+
+	// Verify that results are ingested properly
+	svc.SubmitDistributedQueryResults(ctx, results, map[string]string{})
+
+	// Make sure the result saved to the datastore
+	host, err = ds.AuthenticateHost(nodeKey)
+	require.Nil(t, err)
+
+	// osquery_info
+	assert.Equal(t, "darwin", host.Platform)
+	assert.Equal(t, "1.8.2", host.OsqueryVersion)
+
+	// system_info
+	assert.Equal(t, 17179869184, host.PhysicalMemory)
+	assert.Equal(t, "computer.local", host.HostName)
+	assert.Equal(t, "uuid", host.UUID)
+
+	// os_version
+	assert.Equal(t, "Mac OS X 10.10.6", host.OSVersion)
+
+	// uptime
+	assert.Equal(t, 1730893*time.Second, host.Uptime)
+
+	// osquery_flags
+	assert.Equal(t, uint(10), host.ConfigTLSRefresh)
+	assert.Equal(t, uint(5), host.DistributedInterval)
+	assert.Equal(t, uint(60), host.LoggerTLSPeriod)
+
+	mockClock.AddTime(1 * time.Minute)
+
+	// Now no detail queries should be required
+	host, err = ds.AuthenticateHost(nodeKey)
+	require.Nil(t, err)
+	ctx = hostctx.NewContext(ctx, *host)
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, 0)
+	assert.Zero(t, acc)
+
+	// Advance clock and queries should exist again
+	mockClock.AddTime(1*time.Hour + 1*time.Minute)
+
+	err = svc.SubmitDistributedQueryResults(ctx, kolide.OsqueryDistributedQueryResults{}, map[string]string{})
+	require.Nil(t, err)
+	host, err = ds.AuthenticateHost(nodeKey)
+	require.Nil(t, err)
+
+	ctx = hostctx.NewContext(ctx, *host)
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, len(detailQueries))
+	assert.Zero(t, acc)
+}
+
+func TestDistributedQueries(t *testing.T) {
+	rs := pubsub.NewInmemQueryResults()
+	ds, svc, mockClock := setupOsqueryTests(t)
+	svc.resultStore = rs
+
+	ctx := context.Background()
+
+	nodeKey, err := svc.EnrollAgent(ctx, "", "host123")
+	require.Nil(t, err)
+
+	host, err := ds.AuthenticateHost(nodeKey)
+	require.Nil(t, err)
+
+	host.Platform = "centos"
+	host.HostName = "zwass.local"
+	require.Nil(t, ds.SaveHost(host))
+
+	// Create label
+	n := "foo"
+	q := "select * from foo;"
+	label, err := svc.NewLabel(ctx, kolide.LabelPayload{
+		Name:  &n,
+		Query: &q,
+	})
+	require.Nil(t, err)
+
+	// Record match with label
+	ctx = viewer.NewContext(ctx, viewer.Viewer{
+		User: &kolide.User{
+			ID: 0,
+		},
+	})
+	err = ds.RecordLabelQueryExecutions(host, map[uint]bool{label.ID: true}, mockClock.Now())
+	require.Nil(t, err)
+	err = ds.MarkHostSeen(host, mockClock.Now())
+	require.Nil(t, err)
+	ctx = hostctx.NewContext(ctx, *host)
+
+	q = "select year, month, day, hour, minutes, seconds from time"
+	campaign, err := svc.NewDistributedQueryCampaign(ctx, q, []uint{}, []uint{label.ID})
+	require.Nil(t, err)
+
+	// Manually set the campaign to running (so that it shows up when
+	// requesting queries)
+	campaign.Status = kolide.QueryRunning
+	err = ds.SaveDistributedQueryCampaign(campaign)
+	require.Nil(t, err)
+
+	queryKey := fmt.Sprintf("%s%d", hostDistributedQueryPrefix, campaign.ID)
+
+	// Now we should get the active distributed query
+	queries, acc, err := svc.GetDistributedQueries(ctx)
+	require.Nil(t, err)
+	assert.Len(t, queries, len(detailQueries)+1)
+	assert.Equal(t, q, queries[queryKey])
+	assert.Zero(t, acc)
+
+	expectedRows := []map[string]string{
+		{
+			"year":    "2016",
+			"month":   "11",
+			"day":     "11",
+			"hour":    "6",
+			"minutes": "12",
+			"seconds": "10",
+		},
+	}
+	results := map[string][]map[string]string{
+		queryKey: expectedRows,
+	}
+
+	// TODO use service method
+	readChan, err := rs.ReadChannel(ctx, *campaign)
+	require.Nil(t, err)
+
+	// We need to listen for the result in a separate thread to prevent the
+	// write to the result channel from failing
+	var waitSetup, waitComplete sync.WaitGroup
+	waitSetup.Add(1)
+	waitComplete.Add(1)
+	go func() {
+		waitSetup.Done()
+		select {
+		case val := <-readChan:
+			if res, ok := val.(kolide.DistributedQueryResult); ok {
+				assert.Equal(t, campaign.ID, res.DistributedQueryCampaignID)
+				assert.Equal(t, expectedRows, res.Rows)
+				assert.Equal(t, *host, res.Host)
+			} else {
+				t.Error("Wrong result type")
+			}
+			assert.NotNil(t, val)
+
+		case <-time.After(1 * time.Second):
+			t.Error("No result received")
+		}
+		waitComplete.Done()
+	}()
+
+	waitSetup.Wait()
+	// Sleep a short time to ensure that the above goroutine is blocking on
+	// the channel read (the waitSetup.Wait() is not necessarily sufficient
+	// if there is a context switch immediately after waitSetup.Done() is
+	// called). This should be a small price to pay to prevent flakiness in
+	// this test.
+	time.Sleep(10 * time.Millisecond)
+
+	err = svc.SubmitDistributedQueryResults(ctx, results, map[string]string{})
+	require.Nil(t, err)
+
+	// Now the distributed query should be completed and not returned
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	require.Nil(t, err)
+	assert.Len(t, queries, len(detailQueries))
+	assert.NotContains(t, queries, queryKey)
+	assert.Zero(t, acc)
+
+	waitComplete.Wait()
+}
+
+func TestLabelQueries(t *testing.T) {
+	ds, svc, mockClock := setupOsqueryTests(t)
+	ctx := context.Background()
+
+	_, err := svc.EnrollAgent(ctx, "", "host123")
+	require.Nil(t, err)
+
+	hosts, err := ds.ListHosts(kolide.ListOptions{})
+	require.Nil(t, err)
+	require.Len(t, hosts, 1)
+	host := hosts[0]
+
+	ctx = hostctx.NewContext(ctx, *host)
+
+	// With a new host, we should get the detail queries (and accelerate
+	// should be turned on so that we can quickly fill labels)
+	queries, acc, err := svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, len(detailQueries))
+	assert.NotZero(t, acc)
+
+	// Simulate the detail queries being added
+	host.DetailUpdateTime = mockClock.Now().Add(-1 * time.Minute)
+	host.Platform = "darwin"
+	host.HostName = "zwass.local"
+	ds.SaveHost(host)
+	ctx = hostctx.NewContext(ctx, *host)
+
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, 0)
+	assert.Zero(t, acc)
+
+	labels := []kolide.Label{
+		kolide.Label{
+			Name:     "label1",
+			Query:    "query1",
+			Platform: "darwin",
+		},
+		kolide.Label{
+			Name:     "label2",
+			Query:    "query2",
+			Platform: "darwin",
+		},
+		kolide.Label{
+			Name:     "label3",
+			Query:    "query3",
+			Platform: "darwin,linux",
+		},
+		kolide.Label{
+			Name:     "label4",
+			Query:    "query4",
+			Platform: "linux",
+		},
+	}
+
+	for _, label := range labels {
+		_, err := ds.NewLabel(&label)
+		assert.Nil(t, err)
+	}
+
+	// Now we should get the label queries
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, 3)
+	assert.Zero(t, acc)
+
+	// Record a query execution
+	err = svc.SubmitDistributedQueryResults(
+		ctx,
+		map[string][]map[string]string{
+			hostLabelQueryPrefix + "1": {{"col1": "val1"}},
+		},
+		map[string]string{},
+	)
+	assert.Nil(t, err)
+
+	// Verify that labels are set appropriately
+	hostLabels, err := ds.ListLabelsForHost(host.ID)
+	assert.Len(t, hostLabels, 1)
+	assert.Equal(t, "label1", hostLabels[0].Name)
+
+	// Now that query should not be returned
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, 2)
+	assert.NotContains(t, queries, "kolide_label_query_1")
+	assert.Zero(t, acc)
+
+	// Advance the time
+	mockClock.AddTime(1*time.Hour + 1*time.Minute)
+
+	// Keep the host details fresh
+	host.DetailUpdateTime = mockClock.Now().Add(-1 * time.Minute)
+	ds.SaveHost(host)
+	ctx = hostctx.NewContext(ctx, *host)
+
+	// Now we should get all the label queries again
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, 3)
+	assert.Zero(t, acc)
+
+	// Record a query execution
+	err = svc.SubmitDistributedQueryResults(
+		ctx,
+		map[string][]map[string]string{
+			hostLabelQueryPrefix + "2": {{"col1": "val1"}},
+			hostLabelQueryPrefix + "3": {},
+		},
+		map[string]string{},
+	)
+	assert.Nil(t, err)
+
+	// Now these should no longer show up in the necessary to run queries
+	queries, acc, err = svc.GetDistributedQueries(ctx)
+	assert.Nil(t, err)
+	assert.Len(t, queries, 1)
+	assert.Zero(t, acc)
+
+	// Verify that labels are set appropriately
+	hostLabels, err = ds.ListLabelsForHost(host.ID)
+	assert.Len(t, hostLabels, 2)
+	expectLabelNames := map[string]bool{"label1": true, "label2": true}
+	for _, label := range hostLabels {
+		assert.Contains(t, expectLabelNames, label.Name)
+		delete(expectLabelNames, label.Name)
 	}
 }
